refactor: useCategoryFilter hook
